Handle new problematic enums in Godot 3.2 beta

+++ bindings_generator/src/api.rs
@@ -91,7 +91,7 @@ pub type ConstantValue = i64;
 #[derive(Deserialize, Debug)]
 pub struct Enum {
     pub name: String,
-    pub values: HashMap<String, u32>,
+    pub values: HashMap<String, i64>,
 }
 
 #[derive(Deserialize, Debug)]
